Fix warning.  Always declare integer variables as "int" (or "unsigned") unless you have a real reason to declare it as something else.  There's no benefit to using "uint8_t" here.

+++ stepmania/src/ScreenNetRoom.h
@@ -18,7 +18,7 @@ public:
 private:
 	CString m_name;
 	CString m_description;
-	uint8_t m_state;
+	unsigned int m_state;
 };
 
 class ScreenNetRoom : public ScreenNetSelectBase
